port from utf8 compiler fix

+++ ayavm.h
@@ -74,7 +74,7 @@ public:
 	void func_parse_new(void);
 
 	unsigned int genrand(void);
-	int genrand_int(int n);
+	size_t genrand_int(size_t n);
 
 	unsigned int genrand_sysfunc(void);
 	int genrand_sysfunc_int(int n);
